+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New `status` command that shows which project is currently tracked and for how long (thanks [/u/radarevada](https://www.reddit.com/r/programming/comments/ozxrd4/trackie_is_a_private_daemonless_time_tracker/h84sukr))
 - New `resume` command starts tracking the last tracked project
 - A custom path for the `trackie.json` file can be specified via the `TRACKIE_CONFIG` environment variable.
+- Add possibility to dump the report as JSON via the `--json` flag
 
 ### Changed
 - The `trackie.json` is now saved in `%APPDATA%/trackie/trackie.json` in Windows and `$XDG_DATA_HOME/trackie/trackie.json` on Unix.